Fix tests so they finally pass 100% with tree ACLs, also fix a couple of bugs
* test SQL permission evaluation
* fix parent_pages refresh after deleting ACL
* fix SQL category permission evaluation under SHRINK mode
* fix SD moving after page move
* fix badlink ACL refresh after creating subject page
* fix badlink ACL refresh after moving pages

+++ storage/intraacl-functions-mysql.sql
@@ -66,42 +66,70 @@ begin
       where p2.page_id=_page_id;
 end //
 
-drop procedure if exists /*_*/refresh_all_parents_for_page //
+drop procedure if exists /*_*/refresh_parent_for_page //
 
-create procedure /*_*/refresh_all_parents_for_page(_page_id int unsigned, _page_namespace int, _page_title text)
+create procedure /*_*/refresh_parent_for_page(_page_id int unsigned, _page_namespace int, _page_title text)
 modifies sql data
+begin
+  declare _parent_id int unsigned default null;
+  select get_parent_for_page(_page_namespace, _page_title) into _parent_id;
+  if _parent_id is not null then
+    replace into /*_*/parent_pages (parent_page_id, page_id) values (_parent_id, _page_id);
+  else
+    delete from /*_*/parent_pages where page_id=_page_id;
+  end if;
+end //
+
+drop function if exists /*_*/get_parent_for_page //
+
+create function /*_*/get_parent_for_page(_page_namespace int, _page_title text) returns int unsigned
+reads sql data
 begin
   declare prev int default 0;
   declare pos int default 0;
-  call /*_*/refresh_parent_pages_for_children(_page_id, _page_namespace, _page_title);
+  declare _parent_id int unsigned default null;
   repeat
     set prev = pos;
-    replace into /*_*/parent_pages (parent_page_id, page_id)
-      select p1.page_id, _page_id from /*_*/page p1
+    select p1.page_id from /*_*/page p1
       inner join /*_*/intraacl_rules r1 on r1.pe_type=10 and r1.pe_id=p1.page_id and r1.child_type=6 and r1.child_id=0
       where p1.page_namespace=_page_namespace
-      and p1.page_title = substr(_page_title, 1, length(_page_title)-pos);
-    if row_count() > 0 then
+      and p1.page_title = substr(_page_title, 1, length(_page_title)-pos)
+      into _parent_id;
+    if _parent_id is not null then
       set pos = 0;
     else
       set pos = locate('/', reverse(_page_title), prev+1);
     end if;
   until pos <= 0 end repeat;
+  return _parent_id;
+end //
+
+drop procedure if exists /*_*/refresh_all_parents_for_page //
+
+create procedure /*_*/refresh_all_parents_for_page(_page_id int unsigned, _page_namespace int, _page_title text)
+modifies sql data
+begin
+  if _page_namespace is null or _page_title is null then
+    select page_namespace, page_title from page where page_id=_page_id into _page_namespace, _page_title;
+  end if;
+  call /*_*/refresh_parent_pages_for_children(_page_id, _page_namespace, _page_title);
+  call /*_*/refresh_parent_for_page(_page_id, _page_namespace, _page_title);
 end //
 
 drop procedure if exists /*_*/refresh_parent_pages_for_parent_children //
 
-create procedure /*_*/refresh_parent_pages_for_parent_children(_page_id int unsigned)
+create procedure /*_*/refresh_parent_pages_for_parent_children(_page_id int unsigned, _page_namespace int, _page_title text)
 modifies sql data
 begin
-  declare parent_id int unsigned default 0;
-  declare parent_namespace int default 0;
-  declare parent_title text default '';
-  select p.page_id, p.page_namespace, p.page_title from /*_*/parent_pages pp, /*_*/page p
-    where pp.page_id=_page_id and pp.parent_page_id=p.page_id into parent_id, parent_namespace, parent_title;
-  delete from /*_*/parent_pages where parent_page_id=_page_id;
-  if parent_id > 0 then
-    call /*_*/refresh_parent_pages_for_children(parent_id, parent_namespace, parent_title);
+  declare _parent_id int unsigned default null;
+  if _page_namespace is null or _page_title is null then
+    select page_namespace, page_title from page where page_id=_page_id into _page_namespace, _page_title;
+  end if;
+  select /*_*/get_parent_for_page(_page_namespace, _page_title) into _parent_id;
+  if _parent_id is null then
+    delete from /*_*/parent_pages where parent_page_id=_page_id;
+  else
+    update /*_*/parent_pages set parent_page_id=_parent_id where parent_page_id=_page_id;
   end if;
 end //
 
@@ -281,7 +309,7 @@ begin
   if OLD.page_namespace=14 then
     call /*_*/do_delete_category_closure_catlinks(NULL, OLD.page_id);
   end if;
-  call /*_*/refresh_parent_pages_for_parent_children(OLD.page_id);
+  call /*_*/refresh_parent_pages_for_parent_children(OLD.page_id, OLD.page_namespace, OLD.page_title);
 end //
 
 drop trigger if exists /*_*/update_category_closure_page //
@@ -303,7 +331,7 @@ begin
         on duplicate key update page_id=values(page_id);
     end if;
     -- update parent/child subpage records
-    call /*_*/refresh_parent_pages_for_parent_children(OLD.page_id);
+    call /*_*/refresh_parent_pages_for_parent_children(OLD.page_id, OLD.page_namespace, OLD.page_title);
     call /*_*/refresh_all_parents_for_page(NEW.page_id, NEW.page_namespace, NEW.page_title);
   end if;
 end //
